# Fixeamos bugs del rotate

+++ trunk/Source/XmlParser/AplicacionParser/AplicacionParser/Image.h
@@ -11,8 +11,8 @@ class Image
 	int height;
 	int width;
 	static void putPixel(SDL_Surface* surface, Uint32 pixel, int x, int y);
-	static int xRotatePixel(double radians, int x, int y);
-	static int yRotatePixel(double radians, int x, int y);
+	static double xRotatePixel(double radians, int x, int y);
+	static double yRotatePixel(double radians, int x, int y);
 	int getRotatedHeight(double radians);
 	int getRotatedWidth(double radians);
 	Uint32 getInterpolatedPixel(Uint32 pixelSI, double xSI, double ySI, Uint32 pixelSD, double xSD, double ySD, Uint32 pixelII, double xII, double yII, Uint32 pixelID, double xID, double yID, double xNow, double yNow, const SDL_PixelFormat* newFormat);
